Home page refactored
