+++ src/styles/footer.css
@@ -1,4 +1,5 @@
 .footer {
+    z-index:1000;
     margin-top: auto;
     position: fixed;
     bottom: 0;
