+++ src/styles/topBar.css
@@ -19,3 +19,4 @@
 .top-navbar {
     min-height: 60px;
 }
+
